Button works woot!

+++ CodeFolder/Functions.h
@@ -1,5 +1,5 @@
 #pragma once
 #include <SFML/Graphics.hpp>
 #include <SFML/Audio.hpp>
-bool isAttackButtonPressed(sf::Vector2f clicked);
+bool isAttackButtonPressed(sf::Vector2i clicked);
 void attackButtonLogic();
